pagination
Added loader, smooth scrool to top, eventListener "resize" on window
