+++ src/js/pagination.js
@@ -1,6 +1,8 @@
 import { apiService } from './filmsAPIServise';
-import { refs } from './mainRendering' ;
+import { refs } from './refs' ;
 import renderMarkupFilmsCard from './renderMarkupFilmsCard';
+import { loader } from './loader';
+import buildingPagination from './renderingPagination';
 
 const buttonBack = document.querySelector('.button-back');
 const buttonForward = document.querySelector('.button-forward');
@@ -8,98 +10,6 @@ const buttonsPagesList = document.querySelector('.buttons-pages');
 
 let screenWidth = null;
 let lastPage = null;
-
-// Рендеринг пагінації
-export default function buildingPagination({ page, total_pages: totalPages }) {
-    screenWidth = window.screen.width;
-    lastPage = totalPages;
-    
-// Пагінація для Mobile
-  if (screenWidth < 768) {
-    if (page < 4) {
-      buttonsPagesList.innerHTML = `<li><button class="button-pagination button-page" type="button" id="page-1">1</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-2">2</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-3">3</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-4">4</button></li>
-    <li><button class="button-pagination button-page" type="button id="page-5">5</button></li>
-`;
-        const currentPage = document.querySelector(`#page-${page}`);
-        currentPage.classList.add('current-page')
-        return;
-    }
-
-    if (page + 2 >= totalPages) {
-      buttonsPagesList.innerHTML = `
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages - 4}">${totalPages - 4}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages - 3}">${totalPages - 3}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages - 2}">${totalPages - 2}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages - 1}">${totalPages - 1}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages}">${totalPages}</button></li>
-    `;
-        const currentPage = document.querySelector(`#page-${page}`);
-        currentPage.classList.add('current-page')
-        return;
-    }
-    if (page >= 4) {
-      buttonsPagesList.innerHTML = `<li><button class="button-pagination button-page" type="button">${page - 2}</button></li>
-    <li><button class="button-pagination button-page" type="button">${page - 1}</button></li>
-    <li><button class="button-pagination button-page current-page" type="button">${page}</button></li>
-    <li><button class="button-pagination button-page" type="button">${page + 1}</button></li>
-    <li><button class="button-pagination button-page" type="button">${page + 2}</button></li>
-    `;
-      return;
-    }
-  }
-
-    // Пагінація для Tablet, Desktop
-  if (page < 5) {
-    buttonsPagesList.innerHTML = `<li><button class="button-pagination button-page" type="button" id="page-1">1</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-2">2</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-3">3</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-4">4</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-5">5</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-6">6</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-7">7</button></li>
-    <li><button class="button-pagination button-page" type="button">...</button></li>
-    <li><button class="button-pagination button-page" type="button">${totalPages}</button></li>
-`;
-    const currentPage = document.querySelector(`#page-${page}`);
-    currentPage.classList.add('current-page')
-    return;
-  }
-
-  if (page + 5 >= totalPages) {
-    buttonsPagesList.innerHTML = `<li><button class="button-pagination button-page" type="button">1</button></li>
-    <li><button class="button-pagination button-page" type="button">...</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages-6}">${totalPages - 6}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages-5}">${totalPages - 5}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages-4}">${totalPages - 4}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages-3}">${totalPages - 3}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages-2}">${totalPages - 2}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages-1}">${totalPages - 1}</button></li>
-    <li><button class="button-pagination button-page" type="button" id="page-${totalPages}">${totalPages}</button></li>
-    `;
-      
-    const currentPage = document.querySelector(`#page-${page}`);
-    currentPage.classList.add('current-page')
-    return;
-  }
-
-  if (page >= 5) {
-    buttonsPagesList.innerHTML = `
-        <li><button class="button-pagination button-page" type="button">1</button></li>
-        <li><button class="button-pagination button-page" type="button">...</button></li>
-        <li><button class="button-pagination button-page" type="button">${page - 2}</button></li>
-        <li><button class="button-pagination button-page" type="button">${page - 1}</button></li>
-        <li><button class="button-pagination button-page current-page" type="button">${page}</button></li>
-        <li><button class="button-pagination button-page" type="button">${page + 1}</button></li>
-        <li><button class="button-pagination button-page" type="button">${page + 2}</button></li>
-        <li><button class="button-pagination button-page" type="button">...</button></li>
-        <li><button class="button-pagination button-page" type="button">${totalPages}</button></li>
-    `;
-      return;
-  }
-}
 
 // Слухач на клік по кнопці з номером сторінки
 buttonsPagesList.addEventListener('click', renderingFilmsByPageNumber);
@@ -116,20 +26,28 @@ function renderingFilmsByPageNumber(evt) {
     if (apiService.requestType === 'trending') {
         const fetchPage = evt.target.textContent;
         apiService.page = Number(fetchPage);
+        loader.show();
         apiService.getPopularMovies().then(data => {
             buildingPagination(data);
-            refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results)
+            refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results);
+            try { localStorage.setItem('currentPage', JSON.stringify(data)) } catch (error) {console.log(error.messsge)}
         }).catch(error => error.messsge);
+        loader.hide();
+        window.scrollTo({ top: 0, left: 0, behavior: "smooth" });
         return
     }
 
     if (apiService.requestType === 'search') {
         const fetchPage = evt.target.textContent;
         apiService.page = Number(fetchPage);
+        loader.show();
         apiService.fetchMovieByQuery().then(data => {
             buildingPagination(data);
             refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results)
+            try { localStorage.setItem('currentPage', JSON.stringify(data)) } catch (error) {console.log(error.messsge)}
         }).catch(error => error.messsge);
+        window.scrollTo({ top: 0, left: 0, behavior: "smooth" });
+        loader.hide()
     }
 }
 
@@ -137,20 +55,28 @@ function goBackOnePage() {
     if (apiService.requestType === 'trending') {
         if (apiService.page !== 1) {
             apiService.page -= 1;
+            loader.show();
             apiService.getPopularMovies().then(data => {
                 buildingPagination(data);
-                refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results)
+                refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results);
+                try { localStorage.setItem('currentPage', JSON.stringify(data)) } catch (error) {console.log(error.message)}
             }).catch(error => error.message);
+            loader.hide();
+            window.scrollTo({ top: 0, left: 0, behavior: "smooth" })
         }
         return
     }
     if (apiService.requestType === 'search') {
         if (apiService.page !== 1) {
             apiService.page -= 1;
+            loader.show();
             apiService.fetchMovieByQuery().then(data => {
                 buildingPagination(data);
-                refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results)
+                refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results);
+                try { localStorage.setItem('currentPage', JSON.stringify(data)) } catch (error) {console.log(error.message)}
             }).catch(error => error.message);
+            loader.hide();
+            window.scrollTo({ top: 0, left: 0, behavior: "smooth" })
         }
     }
 }
@@ -159,20 +85,28 @@ function goForwardOnePage() {
     if (apiService.requestType === 'trending') {
         if (apiService.page !== lastPage) {
             apiService.page += 1;
+            loader.show();
             apiService.getPopularMovies().then(data => {
                 buildingPagination(data);
                 refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results);
+                try { localStorage.setItem('currentPage', JSON.stringify(data)) } catch (error) {console.log(error.message)}
             }).catch(error => error.message);
+            loader.hide();
+            window.scrollTo({ top: 0, left: 0, behavior: "smooth" })
         }
     }
 
     if (apiService.requestType === 'search') {
         if (apiService.page !== lastPage) {
             apiService.page += 1;
+            loader.show();
             apiService.fetchMovieByQuery().then(data => {
                 buildingPagination(data);
                 refs.cardContainer.innerHTML = renderMarkupFilmsCard(data.results)
+                try { localStorage.setItem('currentPage', JSON.stringify(data)) } catch (error) {console.log(error.message)}
             }).catch(error => error.message);
+            loader.hide();
+            window.scrollTo({ top: 0, left: 0, behavior: "smooth" })
         }
     }
 }
