+++ src/sass/_pagination.scss
@@ -54,8 +54,8 @@
     transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
 }
 
-.button-back:hover, .button-back:focus,
-.button-forward:hover,.button-forward:focus {
+.button-back:hover,
+.button-forward:hover {
     background-color: #B92F2C;
 }
